@@ -5,57 +5,62 @@
   "version": "0.1.0",
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
   "dependencies": {
-    "@babel/runtime": "^7.13.10",
-    "@fontsource/lora": "^4.1.0",
-    "@fontsource/raleway": "^4.1.0",
-    "babel-plugin-styled-components": "^1.12.0",
+    "@babel/runtime": "^7.23.5",
+    "@fontsource/lora": "^5.0.8",
+    "@fontsource/raleway": "^5.0.15",
+    "babel-plugin-styled-components": "^2.1.4",
     "color-convert": "^2.0.1",
-    "dotenv": "^8.2.0",
-    "embla-carousel": "^4.1.3",
-    "formik": "^2.2.6",
-    "gatsby": "^2.26.1",
+    "dotenv": "^16.3.1",
+    "embla-carousel": "^7.1.0",
+    "embla-carousel-react": "^8.0.0-rc15",
+    "formik": "^2.4.5",
+    "gatsby": "^5.12.11",
+    "gatsby-core-utils": "^4.12.1",
     "gatsby-image": "^2.5.0",
-    "gatsby-plugin-manifest": "^2.6.1",
-    "gatsby-plugin-offline": "^3.4.0",
-    "gatsby-plugin-react-helmet": "^3.4.0",
-    "gatsby-plugin-react-svg": "^3.0.0",
+    "gatsby-link": "^5.12.1",
+    "gatsby-plugin-manifest": "^5.12.3",
+    "gatsby-plugin-offline": "^6.12.3",
+    "gatsby-plugin-postcss": "^6.12.0",
+    "gatsby-plugin-react-helmet": "^6.12.0",
+    "gatsby-plugin-react-svg": "^3.3.0",
     "gatsby-plugin-scroll-reveal": "0.0.7",
-    "gatsby-plugin-sharp": "^2.8.0",
-    "gatsby-plugin-styled-components": "^3.8.0",
-    "gatsby-plugin-stylelint": "^3.2.0",
+    "gatsby-plugin-sharp": "^5.12.3",
+    "gatsby-plugin-styled-components": "^6.12.0",
     "gatsby-plugin-transition-link": "^1.20.5",
-    "gatsby-source-contentful": "^4.4.0",
-    "gatsby-source-filesystem": "^2.5.0",
-    "gatsby-transformer-sharp": "^2.6.0",
-    "gsap": "^3.6.0",
+    "gatsby-react-router-scroll": "^6.12.0",
+    "gatsby-script": "^2.12.0",
+    "gatsby-source-contentful": "^8.12.4",
+    "gatsby-source-filesystem": "^5.12.1",
+    "gatsby-transformer-sharp": "^5.12.3",
+    "gsap": "^3.12.3",
     "intl": "^1.2.5",
     "normalize.css": "^8.0.1",
-    "prop-types": "^15.7.2",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
+    "postcss": "^8.4.31",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-intl": "^5.10.16",
-    "remark": "^13.0.0",
-    "remark-html": "^13.0.1",
-    "remark-preset-lint-markdown-style-guide": "^4.0.0",
-    "styled-components": "^5.2.1"
+    "react-intl": "^6.5.5",
+    "remark": "^15.0.1",
+    "remark-html": "^16.0.1",
+    "remark-preset-lint-markdown-style-guide": "^5.1.3",
+    "styled-components": "^6.1.1"
   },
   "devDependencies": {
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-prettier": "^7.2.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^9.0.0",
     "eslint-loader": "^4.0.2",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.22.0",
-    "gatsby-plugin-eslint": "^2.0.8",
-    "husky": "^4.3.7",
-    "lint-staged": "^10.5.3",
-    "prettier": "^2.2.1",
-    "stylelint": "^13.9.0",
-    "stylelint-config-recommended": "^3.0.0",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
+    "eslint-plugin-react": "^7.33.2",
+    "husky": "^8.0.3",
+    "lint-staged": "^15.1.0",
+    "prettier": "^3.1.0",
+    "stylelint": "^15.11.0",
+    "stylelint-config-recommended": "^13.0.0",
     "stylelint-config-styled-components": "^0.1.1",
     "stylelint-processor-styled-components": "^1.10.0",
-    "stylelint-webpack-plugin": "^2.1.1"
+    "stylelint-webpack-plugin": "^4.1.1"
   },
   "keywords": [
     "gatsby"
