@@ -3,7 +3,7 @@ import React from 'react';
 import { NotificationS } from '../styles/NotificationStyles';
 
 const Notification = ({ type, lifetime, children }) => (
-  <NotificationS type={type} lifetime={lifetime}>
+  <NotificationS $type={type} $lifetime={lifetime}>
     {children}
   </NotificationS>
 );
