@@ -5,11 +5,12 @@ import {
   IntroTextContainerS,
   ListS,
   ListItemS,
+  IntroTextHeaderWrapperS,
 } from '../../styles/IntroTextStyles';
 
 const IntroText = ({ text }) => (
   <IntroTextContainerS>
-    <div
+    <IntroTextHeaderWrapperS
       data-sal="slide-right"
       data-sal-duration="1500"
       data-sal-easing="ease-out-bounce"
@@ -23,7 +24,7 @@ const IntroText = ({ text }) => (
       >
         {text[0]}
       </H2>
-    </div>
+    </IntroTextHeaderWrapperS>
     <ListS
       data-sal="slide-left"
       data-sal-duration="1500"
@@ -32,6 +33,7 @@ const IntroText = ({ text }) => (
       <ListItemS>{text[1]}</ListItemS>
       <ListItemS>{text[2]}</ListItemS>
       <ListItemS>{text[3]}</ListItemS>
+      <ListItemS>{text[4]}</ListItemS>
     </ListS>
   </IntroTextContainerS>
 );
