@@ -2,9 +2,9 @@ import styled from 'styled-components';
 import { media } from './styles-utils';
 
 export const Spacer = styled.div`
-  height: ${p => p.heightMobile || `70px`};
+  height: ${p => p.$heightMobile || `70px`};
   @media (${media.large}) {
-    height: ${p => p.heightPC || `100px`};
+    height: ${p => p.$heightPC || `100px`};
   }
 `;
 
@@ -15,6 +15,6 @@ export const SliderOverlayS = styled.div`
   left: 0;
   width: 100%;
   height: 100vh;
-  opacity: ${p => p.opacity || 0.8};
+  opacity: ${p => p.$opacity || 0.8};
   pointer-events: none;
 `;
