@@ -6,7 +6,7 @@ export const NotificationS = styled.div`
   position: fixed;
   z-index: 999;
   background: ${(p) =>
-    p.type === 'success' ? `${colors.darkRed}dd` : `#000000dd`};
+    p.$type === 'success' ? `${colors.darkRed}dd` : `#000000dd`};
   color: #fff;
   top: 30%;
   left: 10%;
@@ -21,7 +21,7 @@ export const NotificationS = styled.div`
   text-align: center;
   font-size: 1.6rem;
   line-height: 1.3;
-  animation: opacity-in-out ${(p) => p.lifetime} linear;
+  animation: opacity-in-out ${(p) => p.$lifetime} linear;
   @media (${media.large}) {
     top: 40%;
     bottom: 40%;
